--- RolePlay/Personnages.json
@@ -1,1 +1,0 @@
-{"Roxnnis":{"Nom":"Roxnnis","LVL":2,"Money":3827,"Weight":20,"PV":{"Actuel":23,"Total":23},"Stats":{"PHY":5,"MEN":13,"INT":14,"SOC":4,"CHA":5},"Weapons":{"Principale":{"Nom":"Revolver composite","Type":"Pistolet","Hand":1,"Weight":5,"Damage":10,"Precision":40,"Effects":{"Mi-Distance":"+20 de précision à mi-distance d'une cible."},"Munitions":{"Chargeur":6,"Reserve":30,"Recharge":1}}},"Armors":{"Principale":{"Nom":"Toison de sniper","Type":"Cuir","Weight":15,"Res":{"PHY":5,"MEN":5,"OTHERS":{"None":"None"}},"Effects":{"Longue distance":"+30 de précision à longue distance d'une cible."}}},"Inv":{"obj1":{"Nom":"Corde","Quantity":1,"Weight":15,"Remain":15,"Unity":"m"},"obj2":{"Nom":"Corde","Quantity":1,"Weight":15,"Remain":15,"Unity":"m"}},"Dcm":"Aucun"},"Dexhort":{"Nom":"Dexhort","LVL":1,"Money":975,"Weight":0,"PV":{"Actuel":1,"Total":1},"Stats":{"PHY":0,"MEN":0,"INT":0,"SOC":0,"CHA":0},"Weapons":{"Principale":{"Nom":"Aucune","Type":"None","Hand":0,"Weight":0,"Damage":0,"Precision":0}},"Armors":{"Principale":{"Nom":"Aucune","Type":"None","Weight":0,"Res":{"PHY":0,"MEN":0,"OTHERS":{"None":"None"}}}},"Inv":{"Patate":{"Nom":"Patate","Quantity":1,"Stackable":false,"Weight":1},"Patatefalse":{"Nom":"Patate","Quantity":1,"Stackable":false,"Weight":1}},"Dcm":"Aucun"},"Côtelette":{"Nom":"BeafSteak","LVL":2,"Money":0,"Weight":0,"PV":{"Actuel":13,"Total":13},"Stats":{"PHY":0,"MEN":13,"INT":10,"SOC":0,"CHA":12},"Weapons":{"Principale":{"Nom":"Aucune","Type":"None","Hand":0,"Weight":0,"Damage":0,"Precision":0,"Effects":{"None":"Aucun effet"}},"Auxiliaire":{"Nom":"Aucune","Type":"None","Hand":0,"Weight":0,"Damage":0,"Precision":0,"Effects":{"None":"Aucun effet"}}},"Armors":{"Principale":{"Nom":"Aucune","Type":"None","Weight":0,"Res":{"PHY":0,"MEN":0,"OTHERS":{"None":"None"}},"Effects":{"None":"Aucun effet"}}},"Inv":{},"Dcm":"Aucun"}}
