+++ RolePlay/commands/stats.js
@@ -2,7 +2,6 @@ const { SlashCommandBuilder } = require("@discordjs/builders");
 
 //Personnages
 const fs = require("fs");
-var Pers = fs.readFileSync("./Personnages.json");
 const PInf = require("./fonctions/personnageInfo.js");
 const PAdd = require("./fonctions/personnageAdd.js");
