+++ RolePlay/donnees/Personnages.json
@@ -1,0 +1,1 @@
+{"Boo":{"Nom":"Boo","LVL":1,"Money":1000,"Weight":0,"PV":{"Actuel":999,"Total":999},"Stats":{"PHY":0,"MEN":999,"INT":15,"SOC":0,"CHA":42},"Weapons":{"Principale":{"Nom":"Aucune","Type":"None","Hand":0,"Weight":0,"Damage":0,"Precision":0}},"Armors":{"Principale":{"Nom":"Aucune","Type":"None","Weight":0,"Res":{"PHY":0,"MEN":0,"OTHERS":{"None":"None"}}}},"Inv":{},"Dcm":"Aucun"}}
